+++ test/test_helper.exs
@@ -1,1 +1,3 @@
+Code.require_file("support/task_helpers.ex", __DIR__)
+
 ExUnit.start()